:recycle: refactor: auth 수정
